// ui/src/generated/PatchChange.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Entity } from "./Entity";
import type { EntityKey } from "./EntityKey";
import type { ProjectSettings } from "./ProjectSettings";

export type PatchChange = { "type": "Upsert", entity: Entity, } | { "type": "Remove", key: EntityKey, } | { "type": "Settings", settings: ProjectSettings, };